# Reduce allocations during readNativeFrames, leading to ~15-20% performance improvement