@kaxap 2020-12-19
@suyashkumar could you please take a look on this [fork](https://github.com/kaxap/dicom)? We're seeing 5-10x speed improvements due to reduced hot path allocations.