@kaxap 2020-12-20
@suyashkumar Thanks for the feedback. I was thinking about creating PR but I was not sure about the struct change I made, glad we cleared that up.
I'll create PR with `binary<order>.Uintxx` asap.